+++ spigot/spigot-1-20/src/test/java/net/apartium/spigot/visibility/VisibilityManagerTest.java
@@ -1,16 +1,17 @@
-package net.apartium.cocoabeans.spigot;
+package net.apartium.spigot.visibility;
 
 import be.seeseemelk.mockbukkit.entity.PlayerMock;
-import net.apartium.cocoabeans.spigot.visibility.TestPlayerVisibilityController;
 import net.apartium.cocoabeans.spigot.visibility.VisibilityGroup;
 import net.apartium.cocoabeans.spigot.visibility.VisibilityManager;
 import net.apartium.cocoabeans.spigot.visibility.VisibilityPlayerRemoveType;
+import net.apartium.spigot.SpigotTestBase;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class VisibilityManagerTest extends CocoaBeansTestBase {
+public class VisibilityManagerTest extends SpigotTestBase {
 
     private PlayerMock
             ikfir,
@@ -36,12 +37,17 @@ public class VisibilityManagerTest extends CocoaBeansTestBase {
 
     }
 
+    @Override
+    public void initialize() {
+
+    }
+
     @Test
     void someTestName() {
 
         plugin.getLogger();
 
-        visibilityManager = new VisibilityManager(plugin, new TestPlayerVisibilityController());
+        visibilityManager = new VisibilityManager(plugin);
 
         VisibilityGroup group = visibilityManager.getOrCreateGroup("test");
         VisibilityGroup group1 = visibilityManager.getOrCreateGroup("test2");