+++ spigot/spigot-1-8/src/main/java/net/apartium/cocoabeans/spigot/inventory/ItemUtilsHelpers_1_8_R1.java
@@ -1,0 +1,12 @@
+package net.apartium.cocoabeans.spigot.inventory;
+
+import org.bukkit.Material;
+
+public class ItemUtilsHelpers_1_8_R1 implements ItemUtilsHelpers {
+
+    @Override
+    public String name(Material material) {
+        return material.name();
+    }
+
+}