+++ spigot/spigot-1-8/src/main/java/net/apartium/cocoabeans/spigot/spigot/visibility/PlayerVisibilityController_1_8_R1.java
@@ -1,0 +1,19 @@
+package net.apartium.cocoabeans.spigot.spigot.visibility;
+
+import net.apartium.cocoabeans.spigot.visibility.PlayerVisibilityController;
+import org.bukkit.entity.Player;
+import org.bukkit.plugin.java.JavaPlugin;
+
+public class PlayerVisibilityController_1_8_R1 implements PlayerVisibilityController {
+
+    @Override
+    public void showPlayer(JavaPlugin plugin, Player source, Player target) {
+        source.showPlayer(target);
+    }
+
+    @Override
+    public void hidePlayer(JavaPlugin plugin, Player source, Player target) {
+        source.hidePlayer(target);
+    }
+
+}